# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Elevate API",
    "version": "1.0.0",
    "description": "REST API for Elevate, a dark-themed Kanban and planning app. Authenticate via JWT bearer tokens (from /api/v1/auth/login) or API keys (X-API-Key header or ?api_key= query param). Use workspace_id to scope all operations.",
    "contact": {
      "name": "Elevate"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "API v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "Column": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "position": {
            "type": "integer"
          },
          "taskIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "title",
          "position",
          "taskIds"
        ]
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "authorAvatar": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-16T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "text",
          "created_at"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/columns": {
      "post": {
        "tags": [
          "Columns"
        ],
        "summary": "Create column",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspaceId": {
                    "type": "string",
                    "pattern": "^[a-zA-Z0-9-]+$",
                    "example": "trust-and-safety"
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                },
                "required": [
                  "workspaceId",
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Column created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Column"
                }
              }
            }
          }
        }
      }
    },
    "/columns/{id}": {
      "patch": {
        "tags": [
          "Columns"
        ],
        "summary": "Rename column",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                },
                "required": [
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Column"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Columns"
        ],
        "summary": "Delete column",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
    "/columns/reorder": {
      "post": {
        "tags": [
          "Columns"
        ],
        "summary": "Reorder columns",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspaceId": {
                    "type": "string",
                    "pattern": "^[a-zA-Z0-9-]+$",
                    "example": "trust-and-safety"
                  },
                  "columnOrder": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "workspaceId",
                  "columnOrder"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reordered"
          }
        }
      }
    },
    "/columns/archive/{id}": {
      "post": {
        "tags": [
          "Columns"
        ],
        "summary": "Archive column",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Archived"
          }
        }
      }
    },
    "/columns/restore/{id}": {
      "post": {
        "tags": [
          "Columns"
        ],
        "summary": "Restore column",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Restored"
          }
        }
      }
    },
    "/tasks/{taskId}/comments": {
      "post": {
        "tags": [
          "Comments"
        ],
        "summary": "Add comment",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "taskId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 5000
                  }
                },
                "required": [
                  "text"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        }
      }
    },
...
  }
}